requirements.txt
python-dotenv
requests
pandas
yfinance
langchain>=0.2
langchain-openai>=0.1